:truck: Move Git class to VCS/Git
* Hub コマンドも追加したいので、ネームスペースを深くする
